fix(ci): Preserve file permissions in e2e artifacts
The e2e tests were failing on Ubuntu with an EACCES error because the
execute permissions for esbuild were not being preserved when the build
artifacts were uploaded and downloaded.

This change archives the build artifacts into a tarball before uploading
and extracts them in the test jobs. This preserves the file permissions
and fixes the build.